The following changes have been made
(1) Hash tables are no longer used for tracking event_names and their details
(2) We now use *maps* for tracking the event name and details.
(3) Event name and event detail maps are organized at a process level. So, if the region, component category, component and event name, that event is treated as a duplicate
(4) Event name cleanup from maps is still pending.




git-svn-id: https://svn.mcs.anl.gov/repos/cifts/ftb/trunk@736 b452418e-dd38-0410-89fe-f8a55b138f88